very verbose flag
sometimes it's helpful to know the folder being checked,
especially for large directory searches because if
you've got the wrong arguments on a big job, it can take
a long time to find out if you have to wait until the
job completes.

This flag removes a lot of ambiguity and will likely be
used in the setup of a real run, rather than during the
entire run as the volume of print statements slow
execution significantly.